feat(exclude): allow user to exclude events
